refactor: split autoplay.ts into queueHandlers, settingsHandlers, artistHandlers
- Extract queue operations (skip, clear, status) into queueHandlers.ts
- Extract settings operations (mode, genre, analytics) into settingsHandlers.ts
- Extract artist preference operations (prefer, block, list, remove) into artistHandlers.ts
- Refactor main autoplay.ts to thin router pattern with dispatcher
- Maintain all existing functionality and type safety
- Reduce main file from 1074 to ~250 lines
