+++ packages/bot/src/functions/music/commands/autoplay/settingsHandlers.ts
@@ -1,0 +1,438 @@
+import { ChatInputCommandInteraction, ColorResolvable } from 'discord.js'
+import {
+    createEmbed,
+    createErrorEmbed,
+} from '../../../../../../utils/embed/embedHelpers'
+import {
+    EMBED_COLORS,
+    EMOJIS,
+} from '../../../../../../utils/constants/embedConstants'
+import { guildSettingsService } from '../../../../../../services/guild/guildSettingsService'
+import { trackHistoryService } from '../../../../../../services/track/trackHistoryService'
+import { errorLog } from '../../../../../../utils/log/errorLog'
+import { interactionReply } from '../../../../../../utils/interaction/interactionReply'
+
+async function handleAutoplayMode(
+    interaction: ChatInputCommandInteraction,
+): Promise<void> {
+    const guildId = interaction.guildId
+    if (!guildId) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Guild Not Found',
+                        'Unable to retrieve guild information.',
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    const mode = interaction.options.getString('mode')
+
+    if (!mode) {
+        // Get current mode
+        const settings = await guildSettingsService.getGuildSettings(guildId)
+        const currentMode = settings?.autoplayMode ?? 'similar'
+
+        const modeDescriptions: Record<string, string> = {
+            similar: "🎵 Music like what's playing",
+            discover: '🔭 Prioritize new artists',
+            popular: '🔥 Favour your liked tracks',
+        }
+
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createEmbed({
+                        title: '🎚️ Autoplay Mode',
+                        description: `**Current mode:** ${currentMode}\n${modeDescriptions[currentMode] || 'Unknown mode'}`,
+                        color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                        emoji: EMOJIS.AUTOPLAY,
+                        timestamp: true,
+                    }),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    // Set new mode
+    const success = await guildSettingsService.updateGuildSettings(guildId, {
+        autoplayMode: mode as 'similar' | 'discover' | 'popular',
+    })
+
+    if (!success) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Error',
+                        'Failed to update autoplay mode.',
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    const modeEmojis: Record<string, string> = {
+        similar: '🎵',
+        discover: '🔭',
+        popular: '🔥',
+    }
+
+    await interactionReply({
+        interaction,
+        content: {
+            embeds: [
+                createEmbed({
+                    title: '✅ Autoplay mode updated',
+                    description: `${modeEmojis[mode]} Mode set to **${mode}**`,
+                    color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                    emoji: EMOJIS.AUTOPLAY,
+                    timestamp: true,
+                }),
+            ],
+            ephemeral: true,
+        },
+    })
+}
+
+async function handleAutoplayGenreAdd(
+    interaction: ChatInputCommandInteraction,
+): Promise<void> {
+    const guildId = interaction.guildId
+    if (!guildId) return
+
+    const tag = interaction.options.getString('tag')
+    if (!tag) return
+
+    const settings = await guildSettingsService.getGuildSettings(guildId)
+    const genres = settings?.autoplayGenres ?? []
+
+    if (genres.length >= 5) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Limit Reached',
+                        'Maximum 5 genres allowed per guild.',
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    const lowerTag = tag.toLowerCase().trim()
+    if (genres.includes(lowerTag)) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Already Added',
+                        `Genre **${lowerTag}** is already in your list.`,
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    const updated = await guildSettingsService.updateGuildSettings(guildId, {
+        autoplayGenres: [...genres, lowerTag],
+    })
+
+    if (!updated) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [createErrorEmbed('Error', 'Failed to add genre.')],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    await interactionReply({
+        interaction,
+        content: {
+            embeds: [
+                createEmbed({
+                    title: '✅ Genre added',
+                    description: `Added **${lowerTag}** to autoplay genres.`,
+                    color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                    emoji: EMOJIS.AUTOPLAY,
+                    timestamp: true,
+                }),
+            ],
+            ephemeral: true,
+        },
+    })
+}
+
+async function handleAutoplayGenreRemove(
+    interaction: ChatInputCommandInteraction,
+): Promise<void> {
+    const guildId = interaction.guildId
+    if (!guildId) return
+
+    const tag = interaction.options.getString('tag')
+    if (!tag) return
+
+    const settings = await guildSettingsService.getGuildSettings(guildId)
+    const genres = settings?.autoplayGenres ?? []
+
+    const lowerTag = tag.toLowerCase().trim()
+    if (!genres.includes(lowerTag)) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Not Found',
+                        `Genre **${lowerTag}** is not in your list.`,
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    const updated = await guildSettingsService.updateGuildSettings(guildId, {
+        autoplayGenres: genres.filter((g: string) => g !== lowerTag),
+    })
+
+    if (!updated) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [createErrorEmbed('Error', 'Failed to remove genre.')],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    await interactionReply({
+        interaction,
+        content: {
+            embeds: [
+                createEmbed({
+                    title: '✅ Genre removed',
+                    description: `Removed **${lowerTag}** from autoplay genres.`,
+                    color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                    emoji: EMOJIS.AUTOPLAY,
+                    timestamp: true,
+                }),
+            ],
+            ephemeral: true,
+        },
+    })
+}
+
+async function handleAutoplayGenreList(
+    interaction: ChatInputCommandInteraction,
+): Promise<void> {
+    const guildId = interaction.guildId
+    if (!guildId) return
+
+    const settings = await guildSettingsService.getGuildSettings(guildId)
+    const genres = settings?.autoplayGenres ?? []
+
+    const description =
+        genres.length > 0
+            ? genres.map((g: string, i: number) => `${i + 1}. ${g}`).join('\n')
+            : 'No genres configured yet.'
+
+    await interactionReply({
+        interaction,
+        content: {
+            embeds: [
+                createEmbed({
+                    title: '🎵 Autoplay Genres',
+                    description,
+                    color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                    emoji: EMOJIS.AUTOPLAY,
+                    timestamp: true,
+                }),
+            ],
+            ephemeral: true,
+        },
+    })
+}
+
+async function handleAutoplayGenreClear(
+    interaction: ChatInputCommandInteraction,
+): Promise<void> {
+    const guildId = interaction.guildId
+    if (!guildId) return
+
+    const settings = await guildSettingsService.getGuildSettings(guildId)
+    const genres = settings?.autoplayGenres ?? []
+
+    if (genres.length === 0) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [createErrorEmbed('No Genres', 'No genres to clear.')],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    const updated = await guildSettingsService.updateGuildSettings(guildId, {
+        autoplayGenres: [],
+    })
+
+    if (!updated) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [createErrorEmbed('Error', 'Failed to clear genres.')],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    await interactionReply({
+        interaction,
+        content: {
+            embeds: [
+                createEmbed({
+                    title: '✅ Genres cleared',
+                    description: 'All autoplay genres have been removed.',
+                    color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                    emoji: EMOJIS.AUTOPLAY,
+                    timestamp: true,
+                }),
+            ],
+            ephemeral: true,
+        },
+    })
+}
+
+async function handleAutoplayAnalytics(
+    interaction: ChatInputCommandInteraction,
+): Promise<void> {
+    const guildId = interaction.guildId
+    if (!guildId) {
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Guild Not Found',
+                        'Unable to retrieve guild information.',
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+        return
+    }
+
+    try {
+        const stats = await trackHistoryService.getAutoplayStats(guildId, 200)
+
+        if (stats.total === 0) {
+            await interactionReply({
+                interaction,
+                content: {
+                    embeds: [
+                        createEmbed({
+                            title: '📈 Autoplay Analytics',
+                            description: 'No play history yet.',
+                            color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+                            emoji: EMOJIS.AUTOPLAY,
+                            timestamp: true,
+                        }),
+                    ],
+                    ephemeral: true,
+                },
+            })
+            return
+        }
+
+        const artistsText =
+            stats.topAutoplayArtists.length > 0
+                ? stats.topAutoplayArtists
+                      .map(
+                          (a, i) =>
+                              `${i + 1}. ${a.artist} — ${a.count} play${a.count !== 1 ? 's' : ''}`,
+                      )
+                      .join('\n')
+                : 'No autoplay data yet.'
+
+        const analyticsEmbed = createEmbed({
+            title: '📈 Autoplay Analytics',
+            description: `**Recent plays:** ${stats.total} tracks in the last 200\n**Autoplay:** ${stats.autoplayCount} (${stats.autoplayPercent}%)\n\n**Top autoplay artists:**\n${artistsText}`,
+            color: EMBED_COLORS.AUTOPLAY as ColorResolvable,
+            emoji: EMOJIS.AUTOPLAY,
+            timestamp: true,
+        })
+
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [analyticsEmbed],
+                ephemeral: true,
+            },
+        })
+    } catch (error) {
+        errorLog({
+            message: 'Failed to fetch autoplay analytics',
+            error,
+        })
+        await interactionReply({
+            interaction,
+            content: {
+                embeds: [
+                    createErrorEmbed(
+                        'Error',
+                        'Failed to retrieve autoplay analytics.',
+                    ),
+                ],
+                ephemeral: true,
+            },
+        })
+    }
+}
+
+async function handleAutoplayGenre(
+    interaction: ChatInputCommandInteraction,
+    subcommand: string | null,
+): Promise<void> {
+    switch (subcommand) {
+        case 'add':
+            return handleAutoplayGenreAdd(interaction)
+        case 'remove':
+            return handleAutoplayGenreRemove(interaction)
+        case 'list':
+            return handleAutoplayGenreList(interaction)
+        case 'clear':
+            return handleAutoplayGenreClear(interaction)
+    }
+}
+
+export {
+    handleAutoplayMode,
+    handleAutoplayGenre,
+    handleAutoplayAnalytics,
+}
